refa: Remove code duplication by creating the `StrategyBase` class
